+++ memor/template.py
@@ -259,6 +259,19 @@ class PromptTemplate:
         json_str = json.dumps(self.to_json())
         return len(json_str.encode())
 
+    def get_missing_variables(self, context: Optional[Dict[str, Any]] = None) -> List[str]:
+        """
+        Return the list of missing template variables.
+
+        :param context: template context
+        """
+        final_context = {}
+        if self._custom_map is not None:
+            final_context.update(self._custom_map)
+        if context is not None:
+            final_context.update(context)
+        return sorted(set(self.variables) - set(final_context))
+
     def render(self, context: Optional[Dict[str, Any]] = None) -> str:
         """
         Render method.